update command (add) to use a better code structure
what changed?
moved reading stored tasks from:
  - command handler
  - util function used to genereate serial ids
to the function used to create new tasks.

why?
when creating new tasks we first need to read stored tasks
  as they required in two places (which mentioned upove)
but this should to be done without:
  - duplicate code in different places
  - implicit/tight coupling between unrelated components
and atm the best place for this is the function mentioned above.

also moved updating stored tasks with created task to that function.